Rounded average ComboTrade price to 9 decimal places for consistency in results between all modes of operation with double data type.

+++ Platform/TickZoomCommon/Transactions/TransactionPairImpl.cs
@@ -124,9 +124,9 @@ namespace TickZoom.Transactions
 			binary.SetProperties(parameters);
 		}
 		
-		public void ChangeSize(double newSize, double price)
+		public void ChangeSize(int newSize, double price)
 		{
-			binary.ChangeSize(newSize,price);
+			binary.ChangeSize(newSize, price);
 		}
 		
 		public string ToStringHeader()
